Refactor file paths and imports in App.jsx and AppContent.jsx
